requirements.txt
gspread==3.6.0
starlette==0.13.4
oauth2client==4.1.3
uvicorn==0.12.1
aiofiles==0.5.0
jinja2==2.11.2
pyjwt==1.7.1
cloudinary==1.24.0

python-multipart==0.0.5
Rpyc==4.1.5

https://github.com/gbozee/orm/archive/refs/tags/0.1.5.3.zip
alembic==1.0.10
databases[postgresql]
psycopg2-binary==2.8.6